ManagedIdentityCredential cleanups
