+++ sdk/identity/identity/src/credentials/managedIdentityCredential/index.ts
@@ -17,15 +17,15 @@ import { arcMsi } from "./arcMsi";
 import { tokenExchangeMsi } from "./tokenExchangeMsi";
 import { fabricMsi } from "./fabricMsi";
 
-const logger = credentialLogger("ManagedIdentityCredential");
+const credentialName = "ManagedIdentityCredential";
+const logger = credentialLogger(credentialName);
 
 /**
- * Attempts authentication using a managed identity that has been assigned
- * to the deployment environment.  This authentication type works in Azure VMs,
- * App Service and Azure Functions applications, and inside of Azure Cloud Shell.
+ * Attempts authentication using a managed identity available at the deployment environment.
+ * This authentication type works in Azure VMs, App Service instances, Azure Functions applications,
+ * Azure Kubernetes Services, Azure Service Fabric instances and inside of the Azure Cloud Shell.
  *
  * More information about configuring managed identities can be found here:
- *
  * https://docs.microsoft.com/en-us/azure/active-directory/managed-identities-azure-resources/overview
  */
 export class ManagedIdentityCredential implements TokenCredential {
@@ -85,7 +85,7 @@ export class ManagedIdentityCredential implements TokenCredential {
       }
     }
 
-    throw new CredentialUnavailableError("ManagedIdentityCredential - No MSI credential available");
+    throw new CredentialUnavailableError(`${credentialName} - No MSI credential available`);
   }
 
   private async authenticateManagedIdentity(
@@ -94,7 +94,7 @@ export class ManagedIdentityCredential implements TokenCredential {
     getTokenOptions?: GetTokenOptions
   ): Promise<AccessToken | null> {
     const { span, updatedOptions } = createSpan(
-      "ManagedIdentityCredential-authenticateManagedIdentity",
+      `${credentialName}.authenticateManagedIdentity`,
       getTokenOptions
     );
 
@@ -136,7 +136,7 @@ export class ManagedIdentityCredential implements TokenCredential {
   ): Promise<AccessToken> {
     let result: AccessToken | null = null;
 
-    const { span, updatedOptions } = createSpan("ManagedIdentityCredential.getToken", options);
+    const { span, updatedOptions } = createSpan(`${credentialName}.getToken`, options);
 
     try {
       // isEndpointAvailable can be true, false, or null,
@@ -198,7 +198,7 @@ export class ManagedIdentityCredential implements TokenCredential {
       // we can safely assume the credential is unavailable.
       if (err.code === "ENETUNREACH") {
         const error = new CredentialUnavailableError(
-          `ManagedIdentityCredential is unavailable. Network unreachable. Message: ${err.message}`
+          `${credentialName}: Unavailable. Network unreachable. Message: ${err.message}`
         );
 
         logger.getToken.info(formatError(scopes, error));
@@ -209,7 +209,7 @@ export class ManagedIdentityCredential implements TokenCredential {
       // we can safely assume the credential is unavailable.
       if (err.code === "EHOSTUNREACH") {
         const error = new CredentialUnavailableError(
-          `ManagedIdentityCredential is unavailable. No managed identity endpoint found. Message: ${err.message}`
+          `${credentialName}: Unavailable. No managed identity endpoint found. Message: ${err.message}`
         );
 
         logger.getToken.info(formatError(scopes, error));
@@ -220,7 +220,7 @@ export class ManagedIdentityCredential implements TokenCredential {
       // and it means that the endpoint is working, but that no identity is available.
       if (err.statusCode === 400) {
         throw new CredentialUnavailableError(
-          `ManagedIdentityCredential: The managed identity endpoint is indicating there's no available identity. Message: ${err.message}`
+          `${credentialName}: The managed identity endpoint is indicating there's no available identity. Message: ${err.message}`
         );
       }
 
@@ -228,13 +228,13 @@ export class ManagedIdentityCredential implements TokenCredential {
       // This will throw silently during any ChainedTokenCredential.
       if (err.statusCode === undefined) {
         throw new CredentialUnavailableError(
-          `ManagedIdentityCredential authentication failed. Message ${err.message}`
+          `${credentialName}: Authentication failed. Message ${err.message}`
         );
       }
 
       // Any other error should break the chain.
       throw new AuthenticationError(err.statusCode, {
-        error: "ManagedIdentityCredential authentication failed.",
+        error: `${credentialName} authentication failed.`,
         error_description: err.message
       });
     } finally {
