+++ sdk/identity/identity/src/credentials/managedIdentityCredential/utils.ts
@@ -1,15 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { AccessToken, GetTokenOptions } from "@azure/core-auth";
-import { PipelineRequestOptions, createPipelineRequest } from "@azure/core-rest-pipeline";
-import { Agent } from "http";
-import { IdentityClient } from "../../client/identityClient";
 import { DefaultScopeSuffix } from "./constants";
-import { MSIExpiresInParser } from "./models";
 
 /**
- * Most MSIs send requests to the IMDS endpoint, or a similar endpoint. These are GET requests that require sending a `resource` parameter on the query.
+ * Most MSIs send requests to the IMDS endpoint, or a similar endpoint.
+ * These are GET requests that require sending a `resource` parameter on the query.
  * This resource can be derived from the scopes received through the getToken call, as long as only one scope is received.
  * Multiple scopes assume that the resulting token will have access to multiple resources, which won't be the case.
  *
@@ -33,25 +29,4 @@ export function mapScopesToResource(scopes: string | string[]): string | undefin
   }
 
   return scope.substr(0, scope.lastIndexOf(DefaultScopeSuffix));
-}
-
-export async function msiGenericGetToken(
-  identityClient: IdentityClient,
-  requestOptions: PipelineRequestOptions,
-  expiresInParser: MSIExpiresInParser | undefined,
-  getTokenOptions: GetTokenOptions = {},
-  agent?: Agent
-): Promise<AccessToken | null> {
-  const request = createPipelineRequest({
-    abortSignal: getTokenOptions.abortSignal,
-    ...requestOptions,
-    allowInsecureConnection: true
-  });
-
-  if (agent) {
-    request.agent = agent;
-  }
-
-  const tokenResponse = await identityClient.sendTokenRequest(request, expiresInParser);
-  return (tokenResponse && tokenResponse.accessToken) || null;
 }
